move to analysis directory
