+++ mlir/lib/IR/CMakeLists.txt
@@ -32,7 +32,6 @@ add_mlir_library(MLIRIR
   PatternMatch.cpp
   Region.cpp
   RegionKindInterface.cpp
-  SliceWalk.cpp
   SymbolTable.cpp
   TensorEncoding.cpp
   Types.cpp
